limit amount of restaurants shown in categories page

+++ src/routes/categories_routes.py
@@ -9,9 +9,10 @@ from services.category_service import category_service as cat_s
 def show_categories():
     if not user_s.is_admin():
         abort(403)
-    notice = request.args.get("notice")
+    error = request.args.get("error")
+    success = request.args.get("success")
     categories = cat_s.get_categories_and_restaurants()
-    return render_template("categories.html", categories=categories, notice=notice)
+    return render_template("categories.html", categories=categories, error=error, success=success)
 
 
 @app.route("/categories/add", methods=["POST"])
@@ -19,11 +20,15 @@ def add_category():
     if session["csrf_token"] != request.form["csrf_token"] or not user_s.is_admin():
         abort(403)
     name = request.form.get("name")
-    info = cat_s.add_category(name)
-    return redirect(url_for(".show_categories", notice=info))
+    success, info = cat_s.add_category(name)
+    if not success:
+        return redirect(url_for(".show_categories", error=info))
+    return redirect(url_for(".show_categories", success=info))
 
 
 @app.route("/categories/<int:category_id>/delete", methods=["POST"])
 def delete_category(category_id):
+    if session["csrf_token"] != request.form["csrf_token"] or not user_s.is_admin():
+        abort(403)
     cat_s.delete_category(category_id)
     return redirect("/categories")
